cerebrum_excidium/requirements.txt
# Requirements for Cerebrum Excidium
python-nmap
paramiko
requests-tor